Rename Adapters -> Fields, Presenters -> Pages
Rename Adapters -> Fields

- `BaseAdapter` -> `Field::Base`
- `BelongsToAdapter` -> `Field::BelongsTo`
- `EmailAdapter` -> `Field::Email`
- `ImageAdapter` -> `Field::Image`
- `StringAdapter` -> `Field::String`

Rename Presenters -> Pages

- `BasePresenter` -> `Page::Base`
- `ShowPresenter` -> `Page::Show`
- `IndexPresenter` -> `Page::Index`
- `FormPresenter` -> `Page::Form`

Other:

- Rename @presenter -> @page in controller & views
- update README
- Remove duplication between BaseDashboard and Page::Base
- Rename `...Dashboard#field_types` -> attribute_types
