+++ app/views/dashboard/edit.html.erb
@@ -1,6 +1,6 @@
-<h1>Editing <%= @presenter.page_title %></h1>
+<h1>Editing <%= @page.page_title %></h1>
 
 <%= render 'form' %>
 
-<%= link_to 'Show', @presenter.resource %> |
-<%= link_to 'Back', @presenter.index_path %>
+<%= link_to 'Show', @page.resource %> |
+<%= link_to 'Back', @page.index_path %>
