+++ app/views/dashboard/new.html.erb
@@ -1,5 +1,5 @@
-<h1>New <%= @presenter.resource_name.titleize %></h1>
+<h1>New <%= @page.resource_name.titleize %></h1>
 
 <%= render 'form' %>
 
-<%= link_to 'Back', @presenter.index_path %>
+<%= link_to 'Back', @page.index_path %>
