+++ app/views/fields/form/_image.html.erb
