+++ app/views/fields/index/_image.html.erb
