+++ app/views/fields/index/_string.html.erb
