+++ app/views/fields/show/_email.html.erb
