+++ app/views/fields/show/_string.html.erb
