--- lib/adapters/email_adapter.rb
@@ -1,8 +1,0 @@
-require "action_controller/base"
-require_relative "base_adapter"
-
-class EmailAdapter < BaseAdapter
-  def adapter_name
-    :email
-  end
-end
