--- lib/adapters/image_adapter.rb
@@ -1,7 +1,0 @@
-require_relative "./base_adapter"
-
-class ImageAdapter < BaseAdapter
-  def adapter_name
-    :image
-  end
-end
