--- lib/adapters/string_adapter.rb
@@ -1,7 +1,0 @@
-require_relative "base_adapter"
-
-class StringAdapter < BaseAdapter
-  def adapter_name
-    :string
-  end
-end
