+++ lib/fields/email.rb
@@ -1,0 +1,6 @@
+require_relative "base"
+
+module Field
+  class Email < Field::Base
+  end
+end
