+++ lib/fields/image.rb
@@ -1,0 +1,6 @@
+require_relative "base"
+
+module Field
+  class Image < Field::Base
+  end
+end
