+++ lib/fields/string.rb
@@ -1,0 +1,6 @@
+require_relative "base"
+
+module Field
+  class String < Field::Base
+  end
+end
